#14 Refactor: organized the enums into a single package
& cherry-picked the readme from main
& changed the 'same' keyword that is to keep the transfer syntax to 'keep'
