+++ src/main/java/pt/ua/imodec/ImodecPluginSet.java
@@ -8,9 +8,9 @@ import pt.ua.dicoogle.sdk.PluginSet;
 import pt.ua.dicoogle.sdk.StorageInterface;
 import pt.ua.dicoogle.sdk.settings.ConfigurationHolder;
 import pt.ua.imodec.storage.ImodecStoragePlugin;
-import pt.ua.imodec.util.Format;
-import pt.ua.imodec.util.Native;
-import pt.ua.imodec.util.NewFormat;
+import pt.ua.imodec.util.formats.Format;
+import pt.ua.imodec.util.formats.Native;
+import pt.ua.imodec.util.formats.NewFormat;
 import pt.ua.imodec.webservice.ImodecJettyPlugin;
 
 import java.util.Arrays;
