+++ src/main/java/pt/ua/imodec/util/formats/Native.java
@@ -1,4 +1,4 @@
-package pt.ua.imodec.util;
+package pt.ua.imodec.util.formats;
 
 import org.dcm4che2.data.TransferSyntax;
 
@@ -8,7 +8,7 @@ public enum Native implements Format {
     EXPLICIT_VR_BIG_ENDIAN(TransferSyntax.ExplicitVRBigEndian, "ebe"),
     IMPLICIT_VR_LITTLE_ENDIAN(TransferSyntax.ImplicitVRLittleEndian, "ile"),
     EXPLICIT_VR_LITTLE_ENDIAN(TransferSyntax.ExplicitVRLittleEndian, "ele"),
-    UNCHANGED(null, "same");
+    UNCHANGED(null, "keep");
 
     private final TransferSyntax transferSyntax;
     private final String id;
